[#58] feat: 누적 신고 수가 1 이상인 회원 목록 조회

+++ src/main/java/com/web/stard/domain/member/domain/entity/Member.java
@@ -46,7 +46,7 @@ public class Member extends BaseEntity implements UserDetails {
 
     @ColumnDefault("0")
     @Column(name = "report_count")
-    private double reportCount; // 누적 신고 수
+    private int reportCount; // 누적 신고 수
 
     @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JoinColumn(name = "profile_id")
